created dto mappers and moved them to weather service + added city length constraint

+++ src/app/src/main/java/org/weather/app/controller/WeatherController.java
@@ -1,6 +1,7 @@
 package org.weather.app.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -8,68 +9,45 @@ import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.ResourceAccessException;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import org.weather.app.ErrorMessage;
-import org.weather.app.dto.OpenWeatherDataDto;
 import org.weather.app.dto.OpenWeatherGeoDto;
-import org.weather.app.dto.SavedWeatherDto;
 import org.weather.app.model.Location;
 import org.weather.app.model.User;
 import org.weather.app.model.UserSession;
 import org.weather.app.repository.LocationRepository;
-import org.weather.app.service.SessionService;
 import org.weather.app.service.WeatherService;
 
-import java.math.RoundingMode;
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.function.Function;
-import java.util.function.Predicate;
-import java.util.stream.Collectors;
 
 @Controller
 public class WeatherController {
-    private final SessionService sessionService;
+    @Value("${city.max.length}")
+    private int MAX_CITY_LENGTH;
+
     private final LocationRepository locationRepository;
     private final WeatherService weatherService;
 
-    public WeatherController(SessionService sessionService, LocationRepository locationRepository, WeatherService weatherService) {
-        this.sessionService = sessionService;
+    public WeatherController(LocationRepository locationRepository, WeatherService weatherService) {
         this.locationRepository = locationRepository;
         this.weatherService = weatherService;
     }
 
     @GetMapping("/weather")
     public String get(HttpServletRequest request) {
-        UserSession userSession = sessionService.getSessionFromCookie(request.getCookies());
-        if (!sessionService.isSessionValid(userSession)) {
-            return "redirect:/login";
-        }
+        UserSession userSession = (UserSession) request.getAttribute("currentSession");
+
         User currentUser = userSession.getUser();
         List<Location> locations = locationRepository.findAllByUserId(currentUser.getId());
-        request.setAttribute("savedWeathers", locations.stream()
-                .map(loc -> {
-                    SavedWeatherDto dto = new SavedWeatherDto(
-                            loc.getId(),
-                            weatherService.findWeatherByLatAndLon(loc.getLatitude(), loc.getLongitude()));
-                    dto.response.name = loc.getName();
-                    return dto;
-                })
-                .collect(Collectors.toList())
-        );
+
+        request.setAttribute("savedWeathers", weatherService.mapToSavedWeatherDto(locations));
         request.setAttribute("login", currentUser.getLogin());
         return "weather";
     }
 
     @PostMapping("/weather")
     public String search(RedirectAttributes redirectAttributes, HttpServletRequest request) {
-        UserSession userSession = sessionService.getSessionFromCookie(request.getCookies());
-        if (!sessionService.isSessionValid(userSession)) {
-            return "redirect:/login";
-        }
-
         String city = request.getParameter("city");
-        if (city == null || city.isBlank()) {
+        if (city == null || city.isBlank() || city.length() > MAX_CITY_LENGTH) {
             redirectAttributes.addFlashAttribute("errorMessage", ErrorMessage.INVALID_CITY);
             return "redirect:/weather";
         }
@@ -89,34 +67,14 @@ public class WeatherController {
             redirectAttributes.addFlashAttribute("errorMessage", ErrorMessage.CITY_NOT_FOUND);
             return "redirect:/weather";
         }
-        citiesDto.forEach(c -> System.out.println(c.name));
 
-        redirectAttributes.addFlashAttribute("foundWeathers",
-                citiesDto.stream()
-                        .map(dto -> {
-                            OpenWeatherDataDto weather = weatherService.findWeatherByLatAndLon(dto.lat, dto.lon);
-                            weather.name = dto.name;
-                            return weather;
-                        })
-                        .filter(distinctByKey(dto ->
-                                dto.coord.lat.setScale(1, RoundingMode.HALF_UP) + "," +
-                                dto.coord.lon.setScale(1, RoundingMode.HALF_UP)
-                        ))
-                        .collect(Collectors.toList()));
+        redirectAttributes.addFlashAttribute("foundWeathers", weatherService.mapToOpenWeatherDataDto(citiesDto));
         return "redirect:/weather";
-    }
-
-    private static <T> Predicate<T> distinctByKey(Function<T, Object> keyExtractor) {
-        Set<Object> seen = ConcurrentHashMap.newKeySet();
-        return t -> seen.add(keyExtractor.apply(t));
     }
 
     @PostMapping("/saveWeather")
     public String save(RedirectAttributes redirectAttributes, HttpServletRequest request) {
-        UserSession userSession = sessionService.getSessionFromCookie(request.getCookies());
-        if (!sessionService.isSessionValid(userSession)) {
-            return "redirect:/login";
-        }
+        UserSession userSession = (UserSession) request.getAttribute("currentSession");
 
         String latitude = request.getParameter("lat");
         String longitude = request.getParameter("lon");
@@ -133,10 +91,7 @@ public class WeatherController {
 
     @PostMapping("/deleteWeather")
     public String delete(RedirectAttributes redirectAttributes, HttpServletRequest request) {
-        UserSession userSession = sessionService.getSessionFromCookie(request.getCookies());
-        if (!sessionService.isSessionValid(userSession)) {
-            return "redirect:/login";
-        }
+        UserSession userSession = (UserSession) request.getAttribute("currentSession");
 
         String locationId = request.getParameter("locationId");
         if (locationId == null || locationId.isBlank()) {
@@ -150,7 +105,13 @@ public class WeatherController {
             return "redirect:/weather";
         }
 
+        if (!deleteLocation.getUser().getId().equals(userSession.getUser().getId())) {
+            redirectAttributes.addFlashAttribute("errorMessage", ErrorMessage.INVALID_PARAMS);
+            return "redirect:/weather";
+        }
+
         locationRepository.delete(deleteLocation);
         return "redirect:/weather";
     }
+
 }
